+++ ChangeLog.rst
@@ -8,6 +8,8 @@
 * Fixed the way how errors during project checks are processed. Now a
   project with error will be disabled and check will have a traceback in
   it's ``traceback`` slot.
+* Fixed the issue with leaving checked out repositories in the
+  ``/tmp/checking`` directory after the repository was checked for updates.
 
 0.3.1 (2019-02-08)
 ==================